pyproject.toml
[project]
name = "sizebot"
version = "3.8.0"
description = "sizebot"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    { name = "DigiDuncan", email = "digiduncan@gmail.com" },
    { name = "Natalie Fearnley", email = "nfearnley@gmail.com" }
]
classifiers = [
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)"
]
requires-python = ">= 3.12"
dependencies = [
    "colored==1.4.2",
    "discord-py==2.3.2",
    "appdirs==1.4.4",
    "toml==0.10.2",
    "numpy==1.26.4",
    "pynacl==1.5.0",
    "digiformatter==0.5.7.2",
    "python-dateutil==2.9.0.post0",
    "requests==2.31.0",
    "inflect==7.0.0",
    "arrow==1.3.0",
    "validator-collection==1.5.0",
    "flask==3.0.2",
    "jinja2==3.1.3",
    "itsdangerous==2.1.2",
    "werkzeug==3.0.1",
    "pytz==2024.1",
    "pillow==10.2.0",
    "tqdm==4.66.2",
    "aiohttp-requests==0.2.3",
    "asyncstdlib==3.12.1",
    "ndjson==0.3.1",
    "packaging==24.0",
    "sizeroyale==1.0.5",
    "numexpr==2.9.0",
    "mre>=0.12.0",
]

[project.optional-dependencies]
linux = [
    "cysystemd>=1.6.2",
    "pyuwsgi>=2.0.23.post0",
]

[tool.rye]
managed = true
dev-dependencies = [
    "pytest==8.1.1",
    "flake8==7.0.0",
    "autopep8==2.0.4",
    "pytest-asyncio==0.23.5.post1",
    "flake8-annotations==3.1.0",
]

[project.urls]
Homepage = "https://github.com/sizedev/sizebot"
Repository = "https://github.com/sizedev/sizebot.git"
Issues = "https://github.com/sizedev/sizebot/issues"
Changelog = "https://github.com/sizedev/sizebot/blob/master/changelogs/3.8.md"

[project.scripts]
sizebot = "sizebot.main:main"
sizebotapi = "sizebotapi.main:main"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["sizebot"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff.lint]
select = [
    "F",        # Pyflakes
    "E", "W",   # pycodestyle
    "N",        # pep8-naming
    "UP",       # pyupgrade
    "ANN",      # flake8-annotations
    "ASYNC",    # flake8-async
    "FBT",      # flake8-boolean-trap
    "B",        # flake8-bugbear
    "A",        # flake8-builtins
    "COM",      # flake8-commas
    "C4",       # flake8-comprehensions
    "DTZ",      # flake8-datetimez
    "ICN",      # flake8-import-conventions
    "LOG",      # flake8-logging
    "PIE",      # flake8-pie
    "T20",      # flake8-print
    "PYI",      # flake8-pyi
    "PT",       # flake8-pytest-style
    "RSE",      # flake8-raise
    "RET",      # flake8-return
    "SLF",      # flake8-self
    "PTH",      # flake8-use-pathlib
    "FIX",      # flake8-fixme
    "ERA",      # eradicate
    "PGH",      # pygrep-hooks
    "RUF",      # Ruff-specific rules

]
ignore = ["E501", "ANN101", "ANN102", "ANN401", "ANN204", "COM812", "PIE790", "RET504", "RUF001", "RUF013"]

[tool.ruff.lint.pep8-naming]
ignore-names = ["fromJSON", "toJSON"]
